--- frontend/react/frontend/chan-kari/src/index.jsx
@@ -1,23 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import { BrowserRouter } from 'react-router-dom';
-import { App } from './App';
-import 'bootstrap/dist/css/bootstrap.min.css';
-import './index.css';
-
-
-ReactDOM.render(
-
-  //<React.StrictMode>
-  //  <App />
-  //</React.StrictMode>,
-  <BrowserRouter>
-    <App />
-
-  </BrowserRouter>,
-  document.getElementById('root')
-);
-
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
